+++ client/src/services/check.ts
@@ -1,26 +1,52 @@
 import { checkType, IBadReview } from './../components/BadReview/BadReviewControllers';
 import globalAxios, { AxiosInstance } from 'axios';
+import { message } from 'antd';
+
+type routesType = Exclude<checkType, 'No type'>;
+
+const ROUTES: Record<routesType, string> = {
+  'Bad comment': 'badcomment',
+  'Did not check': 'maxscore',
+};
 
 export class CheckService {
   private axios: AxiosInstance;
-  private cache: Record<number, Record<checkType, IBadReview[]>>;
+  private cache: Record<number, Record<routesType, IBadReview[]>>;
 
   constructor() {
     this.axios = globalAxios.create({ baseURL: `/api/` });
     this.cache = {};
   }
 
-  async getBadComments(taskId: number, type: checkType): Promise<IBadReview[]> {
+  async getData(taskId: number, type: checkType) {
+    let dataFromService: IBadReview[] = [];
+    try {
+      switch (type) {
+        case 'Bad comment':
+          dataFromService = await this.getDataFromServer(taskId, type);
+          break;
+        case 'Did not check':
+          dataFromService = await this.getDataFromServer(taskId, type);
+          break;
+        case 'No type':
+          return;
+        default:
+          throw new Error('Unsupported type');
+      }
+    } catch (error) {
+      message.error('Something went wrong');
+    }
+    return dataFromService;
+  }
+
+  private async getDataFromServer(taskId: number, type: routesType) {
     if (this.cache?.[taskId]?.[type]) return this.cache[taskId][type];
-    const result = await this.axios.get(`checks/badcomment/${taskId}`);
-    this.cache = { ...this.cache, [taskId]: { ...this.cache[taskId], [type]: result.data.data } };
+    const result = await this.axios.get(`checks/${ROUTES[type]}/${taskId}`);
+    this.saveToCache(taskId, type, result.data.data);
     return result.data.data;
   }
 
-  async getMaxScoreCheckers(taskId: number, type: checkType): Promise<IBadReview[]> {
-    if (this.cache?.[taskId]?.[type]) return this.cache[taskId][type];
-    const result = await this.axios.get(`checks/maxscore/${taskId}`);
-    this.cache = { ...this.cache, [taskId]: { ...this.cache[taskId], [type]: result.data.data } };
-    return result.data.data;
+  private saveToCache(taskId: number, type: routesType, data: IBadReview) {
+    this.cache = { ...this.cache, [taskId]: { ...this.cache[taskId], [type]: data } };
   }
 }
